scripts/requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
idna==3.11
lxml==6.0.2
python-dateutil==2.9.0.post0
requests==2.33.0
six==1.17.0
urllib3==2.6.3
webdav==1.1.7
webdavclient3==3.14.7
